Move content to new header.py. Change time/date params of csv_to_wfdb
