+++ wfdb/io/_header.py
@@ -1,13 +1,11 @@
 import datetime
-import re
-from typing import List, Tuple
 
 import numpy as np
 import pandas as pd
 
 from wfdb.io import _signal
 from wfdb.io import util
-
+from wfdb.io.header import HeaderSyntaxError, rx_record, rx_segment, rx_signal
 
 """
 Notes
@@ -31,12 +29,6 @@ The read vs write default values are different for 2 reasons:
    important fields. When writing, we want to fill in defaults
    so that the user doesn't need to. But when reading, it should
    be clear that the fields are missing.
-
-If all of the fields were filled out in a WFDB header file, they would appear
-in this order with these seperators:
-
-RECORD_NAME/NUM_SEG NUM_SIG SAMP_FREQ/COUNT_FREQ(BASE_COUNT_VAL) SAMPS_PER_SIG BASE_TIME BASE_DATE
-FILE_NAME FORMATxSAMP_PER_FRAME:SKEW+BYTE_OFFSET ADC_GAIN(BASELINE)/UNITS ADC_RES ADC_ZERO CHECKSUM BLOCK_SIZE DESCRIPTION
 
 """
 int_types = (int, np.int64, np.int32, np.int16, np.int8)
@@ -134,53 +126,6 @@ SEGMENT_SPECS = pd.DataFrame(
 
 # Specifications of all WFDB header fields, except for comments
 FIELD_SPECS = pd.concat((RECORD_SPECS, SIGNAL_SPECS, SEGMENT_SPECS))
-
-# Regexp objects for reading headers
-# Record line
-_rx_record = re.compile(
-    r"""
-    [ \t]* (?P<record_name>[-\w]+)
-           /?(?P<n_seg>\d*)
-    [ \t]+ (?P<n_sig>\d+)
-    [ \t]* (?P<fs>\d*\.?\d*)
-           /*(?P<counter_freq>-?\d*\.?\d*)
-           \(?(?P<base_counter>-?\d*\.?\d*)\)?
-    [ \t]* (?P<sig_len>\d*)
-    [ \t]* (?P<base_time>\d{,2}:?\d{,2}:?\d{,2}\.?\d{,6})
-    [ \t]* (?P<base_date>\d{,2}/?\d{,2}/?\d{,4})
-    """,
-    re.VERBOSE,
-)
-
-# Signal line
-_rx_signal = re.compile(
-    r"""
-    [ \t]* (?P<file_name>~?[-\w]*\.?[\w]*)
-    [ \t]+ (?P<fmt>\d+)
-           x?(?P<samps_per_frame>\d*)
-           :?(?P<skew>\d*)
-           \+?(?P<byte_offset>\d*)
-    [ \t]* (?P<adc_gain>-?\d*\.?\d*e?[\+-]?\d*)
-           \(?(?P<baseline>-?\d*)\)?
-           /?(?P<units>[\w\^\-\?%\/]*)
-    [ \t]* (?P<adc_res>\d*)
-    [ \t]* (?P<adc_zero>-?\d*)
-    [ \t]* (?P<init_value>-?\d*)
-    [ \t]* (?P<checksum>-?\d*)
-    [ \t]* (?P<block_size>\d*)
-    [ \t]* (?P<sig_name>[\S]?[^\t\n\r\f\v]*)
-    """,
-    re.VERBOSE,
-)
-
-# Segment line
-_rx_segment = re.compile(
-    r"""
-    [ \t]* (?P<seg_name>[-\w]*~?)
-    [ \t]+ (?P<seg_len>\d+)
-    """,
-    re.VERBOSE,
-)
 
 
 class BaseHeaderMixin(object):
@@ -905,37 +850,6 @@ def wfdb_strptime(time_string: str) -> datetime.time:
     return datetime.datetime.strptime(time_string, time_fmt).time()
 
 
-def parse_header_content(
-    header_content: str,
-) -> Tuple[List[str], List[str]]:
-    """
-    Parse the text of a header file.
-
-    Parameters
-    ----------
-    header_content: str
-        The string content of the full header file
-
-    Returns
-    -------
-    header_lines : List[str]
-        A list of all the non-comment lines
-    comment_lines : List[str]
-        A list of all the comment lines
-    """
-    header_lines, comment_lines = [], []
-    for line in header_content.splitlines():
-        line = line.strip()
-        # Comment line
-        if line.startswith("#"):
-            comment_lines.append(line)
-        # Non-empty non-comment line = header line.
-        elif line:
-            header_lines.append(line)
-
-    return header_lines, comment_lines
-
-
 def _parse_record_line(record_line: str) -> dict:
     """
     Extract fields from a record line string into a dictionary.
@@ -955,7 +869,7 @@ def _parse_record_line(record_line: str) -> dict:
     record_fields = {}
 
     # Read string fields from record line
-    match = _rx_record.match(record_line)
+    match = rx_record.match(record_line)
     if match is None:
         raise HeaderSyntaxError("invalid syntax in record line")
     (
@@ -1031,7 +945,7 @@ def _parse_signal_lines(signal_lines):
 
     # Read string fields from signal line
     for ch in range(n_sig):
-        match = _rx_signal.match(signal_lines[ch])
+        match = rx_signal.match(signal_lines[ch])
         if match is None:
             raise HeaderSyntaxError("invalid syntax in signal line")
         (
@@ -1105,7 +1019,7 @@ def _read_segment_lines(segment_lines):
 
     # Read string fields from signal line
     for i in range(len(segment_lines)):
-        match = _rx_segment.match(segment_lines[i])
+        match = rx_segment.match(segment_lines[i])
         if match is None:
             raise HeaderSyntaxError("invalid syntax in segment line")
         (
@@ -1118,7 +1032,3 @@ def _read_segment_lines(segment_lines):
             segment_fields["seg_len"][i] = int(segment_fields["seg_len"][i])
 
     return segment_fields
-
-
-class HeaderSyntaxError(ValueError):
-    """Invalid syntax found in a WFDB header file."""
